requirements.txt
pandas
streamlit==1.33.0
streamlit-server-state==0.17.1
langchain==0.1.16
langchain-community==0.0.34
langchain-core==0.1.50
langchain-openai==0.1.1
langsmith==0.1.52
qdrant-client==1.8.2
sentence-transformers==2.6.1
langchain-mongodb==0.1.6
certifi==2024.7.4
pypdf==4.3.0
injector==0.22.0
pyyaml==6.0.1
fastapi==0.111.1
uvicorn==0.30.3